raftstore: move structs to mod unsafe_recovery
