+++ components/raftstore/src/store/mod.rs
@@ -57,11 +57,8 @@ pub use self::{
     },
     peer::{
         can_amend_read, get_sync_log_from_request, make_transfer_leader_response,
-        propose_read_index, should_renew_lease, ForceLeaderState, Peer, PeerStat, ProposalContext,
-        ProposalQueue, RequestInspector, RequestPolicy, SnapshotRecoveryWaitApplySyncer,
-        UnsafeRecoveryExecutePlanSyncer, UnsafeRecoveryFillOutReportSyncer,
-        UnsafeRecoveryForceLeaderSyncer, UnsafeRecoveryState, UnsafeRecoveryWaitApplySyncer,
-        TRANSFER_LEADER_COMMAND_REPLY_CTX,
+        propose_read_index, should_renew_lease, Peer, PeerStat, ProposalContext, ProposalQueue,
+        RequestInspector, RequestPolicy, TRANSFER_LEADER_COMMAND_REPLY_CTX,
     },
     peer_storage::{
         clear_meta, do_snapshot, write_initial_apply_state, write_initial_raft_state,
@@ -79,7 +76,11 @@ pub use self::{
     },
     transport::{CasualRouter, ProposalRouter, SignificantRouter, StoreRouter, Transport},
     txn_ext::{LocksStatus, PeerPessimisticLocks, PessimisticLockPair, TxnExt},
-    unsafe_recovery::UnsafeRecoveryHandle,
+    unsafe_recovery::{
+        ForceLeaderState, SnapshotRecoveryWaitApplySyncer, UnsafeRecoveryExecutePlanSyncer,
+        UnsafeRecoveryFillOutReportSyncer, UnsafeRecoveryForceLeaderSyncer, UnsafeRecoveryHandle,
+        UnsafeRecoveryState, UnsafeRecoveryWaitApplySyncer,
+    },
     util::{RegionReadProgress, RegionReadProgressRegistry},
     worker::{
         metrics as worker_metrics, AutoSplitController, BatchComponent, Bucket, BucketRange,
