+++ components/raftstore/src/store/msg.rs
@@ -31,7 +31,7 @@ use super::{local_metrics::TimeTracker, region_meta::RegionMeta, FetchedLogs, Re
 use crate::store::{
     fsm::apply::{CatchUpLogs, ChangeObserver, TaskRes as ApplyTaskRes},
     metrics::RaftEventDurationType,
-    peer::{
+    unsafe_recovery::{
         SnapshotRecoveryWaitApplySyncer, UnsafeRecoveryExecutePlanSyncer,
         UnsafeRecoveryFillOutReportSyncer, UnsafeRecoveryForceLeaderSyncer,
         UnsafeRecoveryWaitApplySyncer,
